chore: merge two khayla folders into one + add styling to second page
