+++ lesson_25/khayla/src/main.tsx
@@ -1,19 +1,12 @@
 import App from './App.tsx';
-import { AddProgram } from './pages/AddProgram/AddProgram.tsx';
-import { Home } from './pages/Home/Home.tsx';
-import { ErrorPage } from './pages/errorElement/ErrorPage.tsx';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import {Home} from './pages/Home/Home.tsx';
+import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { RouterProvider, createBrowserRouter } from 'react-router-dom';
-
-
+import {RouterProvider, createBrowserRouter} from 'react-router-dom';
 
 import './index.scss';
-
-
-
-
+import { SecondPage } from './pages/SecondPage/SecondPage.tsx';
 
 const queryClient = new QueryClient();
 
@@ -27,13 +20,9 @@ const router = createBrowserRouter([
         element: <Home />,
       },
       {
-        path: '/AddProgram',
-        element: <AddProgram />,
-      },
-      {
-        path: '/errorElement',
-        element: < ErrorPage/>,
-      },
+        path: '/second-page',
+        element: <SecondPage />,
+      }
     ],
   },
 ]);
