--- lesson_25/khayla/src/pages/AddProgram/AddProgram.scss
@@ -1,3 +1,0 @@
-body{
-    align-items: center;
-}
