+++ lesson_25/khayla/src/pages/SecondPage/SecondPage.scss
@@ -1,0 +1,58 @@
+.program-form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 60px 20px;
+    max-width: 600px;
+    margin: 0 auto;
+    background-color: #f5f5f5;
+    border-radius: 12px;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+  }
+  
+  .form-title {
+    text-align: center;
+    font-size: 2.5rem;
+    font-family: Montserrat, sans-serif;
+    font-weight: 900;
+    color: #444;
+    margin-top: 60px;
+    margin-bottom: 30px;
+  }
+  
+  .input-field {
+    width: 100%;
+    padding: 0.75rem 1rem;
+    margin-bottom: 20px;
+    font-size: 1rem;
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    font-family: Poppins, sans-serif;
+  }
+  
+  .submit-button {
+    background-color: #f47d26;
+    color: white;
+    border: none;
+    padding: 0.75rem 1.5rem;
+    font-size: 1rem;
+    font-family: Montserrat, sans-serif;
+    border-radius: 8px;
+    cursor: pointer;
+    margin-bottom: 20px;
+    transition: background-color 0.3s ease;
+  }
+  
+  .submit-button:hover {
+    background-color: #d8671f;
+  }
+  
+  .back-button {
+    color: #f47d26;
+    font-family: Montserrat, sans-serif;
+    font-weight: bold;
+    text-decoration: underline;
+    cursor: pointer;
+    font-size: 0.95rem;
+  }
+
