--- lesson_25/khayla/src/pages/errorElement/ErrorPage.tsx
@@ -1,3 +1,0 @@
-export function ErrorPage() {
-    return <>Whoops! Please return to main page</>
-}
