--- lesson_25/khaylasaunders/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "lesson_25",
-  "version": "1.0.0",
-  "description": "## Pre-work",
-  "main": "index.js",
-  "scripts": {
-    "dev": "nodemon server.js",
-    "start": "node server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-
-}
