--- lesson_25/khaylasaunders/template/eslint.config.js
@@ -1,4 +1,0 @@
-// .eslintrc.js
-import eslintConfigPrettier from 'eslint-config-prettier';
-
-export default [eslintConfigPrettier];
