--- lesson_25/khaylasaunders/template/src/components/footer/Footer.scss
@@ -1,9 +1,0 @@
-footer {
-  display: flex;
-  flex-direction: column;
-  padding: 45px 0 70px 0;
-  font-size: 0.8em;
-  color: #3c3c3c;
-  align-items: center;
-  box-shadow: inset 0 5px 12px 4px #e5e5e5;
-}
