--- lesson_25/khaylasaunders/template/src/components/footer/Footer.tsx
@@ -1,6 +1,0 @@
-import './Footer.scss';
-import React from 'react';
-
-export const Footer: React.FC = () => {
-  return <footer className="footer">&copy; 2024 Code Differently</footer>;
-};
