--- lesson_25/khaylasaunders/template/src/pages/Home/Home.scss
@@ -1,82 +1,0 @@
-.hero-section {
-  width: 100%;
-  overflow: hidden;
-  position: relative;
-  background-image: url(hero.jpg);
-  background-attachment: fixed;
-  background-size: cover;
-  padding: 100px 0;
-
-  .hero-overlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.25);
-  }
-
-  .hero-content {
-    color: white;
-    width: 70%;
-    position: relative;
-
-    h2 {
-      color: white;
-    }
-  }
-
-  .hero-title {
-    font-size: 3em;
-    font-weight: 700;
-    margin-bottom: 20px;
-    margin-left: 18%;
-    line-height: 1.2em;
-  }
-
-  .hero-text {
-    background-color: #243e90;
-    padding: 40px 0;
-    padding-left: 18%;
-  }
-
-  .hero-text,
-  .hero-title {
-    padding-right: 14px;
-  }
-}
-
-article {
-  background-color: #f5f5f5;
-  justify-content: center;
-  display: flex;
-  flex-direction: column;
-}
-
-.programs-section {
-  padding: 100px;
-  display: flex;
-  flex-direction: column;
-  max-width: 1280px;
-  margin: auto;
-
-  .programs {
-    display: flex;
-    flex-flow: row wrap;
-    list-style: none;
-    padding: 0;
-    justify-content: space-between;
-  }
-}
-
-.program {
-  display: flex;
-  flex-direction: column;
-  max-width: 48%;
-}
-
-@media screen and (max-width: 1080px) {
-  .program {
-    min-width: 100%;
-  }
-}
