--- lesson_25/khaylasaunders/template/src/pages/Home/index.js
@@ -1,1 +1,0 @@
-export {Home} from './Home';
