--- lesson_25/khaylasaunders/template/src/pages/SecondPage/SecondPage.scss
@@ -1,3 +1,0 @@
-body{
-    align-items: center;
-}
