--- lesson_25/khaylasaunders/template/src/pages/SecondPage/index.js
@@ -1,1 +1,0 @@
-export {SecondPage} from './SecondPage';
