--- lesson_25/khaylasaunders/template/vite.config.ts
@@ -1,19 +1,0 @@
-import react from '@vitejs/plugin-react';
-import {URL, fileURLToPath} from 'url';
-import {defineConfig} from 'vite';
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react()],
-  resolve: {
-    alias: [
-      {
-        find: '@',
-        replacement: fileURLToPath(new URL('./src', import.meta.url)),
-      },
-    ],
-  },
-  server: {
-    host: '0.0.0.0',
-  },
-});
